made class names consistent
